+++ TenjinSDK.xcframework/ios-arm64_x86_64-simulator/TenjinSDK.framework/Headers/TenjinSDK.h
@@ -7,6 +7,7 @@
 #import <Foundation/Foundation.h>
 #import <StoreKit/StoreKit.h>
 
+@class TenjinPurchasesManager;
 @class TJNUserProfileData;
 
 @interface TenjinSDK : NSObject
@@ -81,6 +82,9 @@ andDeferredDeeplink:(NSURL *)url
 //returns the shared Tenjin SDK instance
 + (TenjinSDK *)sharedInstance;
 
+//returns the shared Purchases Manager instance for StoreKit 2 on-demand tracking
++ (TenjinPurchasesManager *)purchasesManager;
+
 #pragma mark - Functionality
 
 //use connect to send connect call. sharedInstanceWithToken automatically does a connect
@@ -134,6 +138,15 @@ andDeferredDeeplink:(NSURL *)url
 
 //notify Tenjin of a new subscription purchase
 - (void)handleSubscriptionPurchase:(SKPaymentTransaction *)transaction;
+
+//track subscription with manual parameters (useful for plugins like Flutter, Unity, etc.)
++ (void)subscriptionWithProductName:(NSString *)productName
+                   andCurrencyCode:(NSString *)currencyCode
+                      andUnitPrice:(NSDecimalNumber *)price
+                  andTransactionId:(NSString *)transactionId
+         andOriginalTransactionId:(NSString *)originalTransactionId
+                  andBase64Receipt:(NSString *)receipt
+                 andSKTransaction:(NSString *)skTransaction;
 
 // GDPR opt-out
 + (void)optOut;
@@ -211,6 +224,8 @@ andDeferredDeeplink:(NSURL *)url
 + (NSString *)sdkVersion;
 
 + (void)setWrapperVersion:(NSString *)wrapperVersion;
+
++ (void)setPluginVersion:(NSString *)plugin version:(NSString *)version;
 
 + (void)setValue:(NSString *)value
           forKey:(NSString *)key;